Some updates to the AuthEngineClass buffer.
There is still an issue for some Windows 7 & 8 users logging in and the RealmServer crashing over it.

Focused issue is this error:

A call to PInvoke function 'mangosVB.RealmServer!mangosVB.RealmServer.AuthEngineClass::BN_bin2bn' has unbalanced the stack. This is likely because the managed PInvoke signature does not match the unmanaged target signature. Check that the calling convention and parameters of the PInvoke signature match the target unmanaged signature.

Line 149 in AuthEngineClass.vb

Code:
Me.BNx = AuthEngineClass.BN_bin2bn(buffer3, buffer3.Length, IntPtr.Zero)

THIS IS HIGH PRIORITY